refactor(p2): extract Renderers into sibling page (stage 2 of 4)
Moves 5 assistant-message rendering Composables/helpers from
ChatMessage.kt into new sibling ChatMessageRenderers.kt (same package):
VirtualizedAssistantText + AssistantMarkdownBlockOrWidgets +
ReasoningWidgetRescue + reasoningWidgetSource + AssistantBubbleSegment.

Byte-equal: bodies and signatures unchanged.
Visibility:
- VirtualizedAssistantText: private -> internal (Stage 3 CMVI Content)
- AssistantMarkdownBlockOrWidgets: private -> internal (Stage 4 MessagePartsBlock)
- ReasoningWidgetRescue: private -> internal (Stage 4 MessagePartsBlock)
- reasoningWidgetSource: unchanged (internal, byte-equal default)
- AssistantBubbleSegment: unchanged (private, only same-sibling caller VAT)

Coordinator strip: 1504 -> 1219 (-285 with import cleanup).
Imports trimmed: 3 orphans (Stroke, GenerativeWidgetSegment, MarkdownTopLevelBlock).
Pre-existing dead preserved: getValue, setValue, fastForEach, height
(layout.height grep-pseudo-alive via Size constructor named arg + DrawScope.size.height
but 0 real Modifier.height(...) usage in baseline — playbook lesson #4 detection).

Baseline 3/467 unit tests preserved.
